Line polygon/box clipping: degeneracies
Make a start with dealing with obvious degeneracies. Current behavior is to count intersection in a "greedy" manner.

This means that if a line intersects an edge exactly, it will count both cells as touched. This seems like an okay choice: afterwards, these duplicates edges can be identified relatively easily, if need be. All alternatives seem worse: by NOT counting lines on edges, we could miss lines running along (internal) cell boundaries entirely.

There is also no clear logic possible(?) to count one, but not the other. We could e.g. only count the counter clockwise edges and not the clockwise ones. This does not seem like an intuitive choice compared to "if a line is located exactly on an edge, both cells are returned".